write to db based on zipcode collection with added onMarket and lastUpdated variables

+++ zillow_rent.py
@@ -6,7 +6,11 @@ import argparse
 import json
 import pymongo
 from pymongo import MongoClient
+from datetime import datetime
 
+cluster = MongoClient("mongodb+srv://aagunawan:dedeku88@cluster0.qfiad.mongodb.net/<dbname>?retryWrites=true&w=majority")
+db = cluster["rental"]
+todays_date = datetime.today().strftime('%Y-%m-%d')
 
 def clean(text):
     if text:
@@ -27,56 +31,49 @@ def get_headers():
 
 def create_url(zipcode, filter):
     # Creating Zillow URL based on the filter.
-    # url = "https://www.zillow.com/homes/53726_rb/"
-   
-    # if filter == "newest":
-    #     url = "https://www.zillow.com/homes/for_sale/{0}/0_singlestory/days_sort".format(zipcode)
-    # elif filter == "cheapest":
-    #     url = "https://www.zillow.com/homes/for_sale/{0}/0_singlestory/pricea_sort/".format(zipcode)
-    # else:
-    # url = "https://www.zillow.com/homes/hillsboro-or-{0}/rent-houses/".format(zipcode)
     url = "https://www.zillow.com/homes/for_rent/{0}_rb/?fromHomePage=true&shouldFireSellPageImplicitClaimGA=false&fromHomePageTab=buy".format(zipcode)
-    # url = "https://www.zillow.com/homes/for_rent/55454_rb/?fromHomePage=true&shouldFireSellPageImplicitClaimGA=false&fromHomePageTab=buy"
+
     print(url)
     return url
 
+def update_days(collection, entry):
+    query = {"zid": entry.get("zid")}
+    updated_value = {"$set": {"daysOnZillow": entry.get("daysOnZillow"), "lastUpdated": todays_date}}
 
-def save_to_file(response):
-    # saving response to `response.html`
+    if (collection.find_one(query)): # if found update daysOnZillow and lastUpdated
+        collection.update_one(query, updated_value)
+    else: # if not found, insert it 
+        collection.insert_one(entry)
+    
+    # look at all houses in current zipcode that do not get updated to latest date and change it to off market
+    query_off_market = {"lastUpdated": {"$ne": todays_date}}
+    updated_value_off_market = {"$set": {"onMarket": False}}
+    collection.update_many(query_off_market, updated_value_off_market)
 
-    with open("response.html", 'w') as fp:
-        fp.write(response.text)
+    return None
 
+def update_db(data, zipcode):
+    collection = db[zipcode] 
 
-def write_data_to_csv(data):
-    # saving scraped data to csv.
-
-    with open("properties-%s.csv" % (zipcode), 'wb') as csvfile:
-        fieldnames = ['title', 'address', 'city', 'state', 'postal_code', 'price', 'bed', 'bath', 'area', 'daysOnZillow', 'url']
-        writer = csv.DictWriter(csvfile, fieldnames=fieldnames)
-        writer.writeheader()
-        for row in data:
-            writer.writerow(row)
-
-def write_to_db(data, zipcode):
-    cluster = MongoClient("mongodb+srv://aagunawan:dedeku88@cluster0.qfiad.mongodb.net/<dbname>?retryWrites=true&w=majority")
-    db = cluster["rental"]
-    collection = db["listings"] 
-    collection.drop()
-    collection.insert_many(data)
+    if collection.count_documents({}) == 0: 
+        collection.drop()
+        collection.insert_many(data)
+    
+    else:
+        for entry in data:
+            update_days(collection, entry)
+    
+    return None
 
 def get_response(url):
+    
     # Getting response from zillow.com.
-
     for i in range(5):
         response = requests.get(url, headers=get_headers())
         print("status code received:", response.status_code)
         if response.status_code != 200:
-            # saving response to file for debugging purpose.
-            save_to_file(response)
             continue
         else:
-            save_to_file(response)
             return response
     return None
 
@@ -88,18 +85,13 @@ def get_data_from_json(raw_json_data):
 
     try:
         json_data = json.loads(cleaned_data)
-        # search_results = json_data.get('searchResults').get('listResults', [])
         search_results = json_data.get('cat1').get('searchResults').get('listResults', [])
-        # print(search_results[0])
-        # print(search_results[1])
         for properties in search_results:
             address = properties.get('addressStreet')
+            zid = properties.get('zpid')
             property_info = properties.get('hdpData', {}).get('homeInfo')
             if (property_info):
-
                 city = property_info.get('city')
-                # print(property_info.get('daysOnZillow'))
-                # print(properties.get('statusType'))
                 state = property_info.get('state')
                 postal_code = property_info.get('zipcode')
                 price = properties.get('price')
@@ -108,10 +100,12 @@ def get_data_from_json(raw_json_data):
                 area = properties.get('area')
                 info = f'{bedrooms} bds, {bathrooms} ba ,{area} sqft'
                 daysOnZillow = property_info.get('daysOnZillow')
+                lastUpdated = todays_date
                 property_url = properties.get('detailUrl')
                 title = properties.get('statusText')
 
-                data = {'address': address,
+                data = {'zid': zid,
+                        'address': address,
                         'city': city,
                         'state': state,
                         'postal_code': postal_code,
@@ -119,8 +113,9 @@ def get_data_from_json(raw_json_data):
                         'bed': bedrooms,
                         'bath': bathrooms,
                         'area': area,
-                        # 'facts and features': info,
                         'daysOnZillow': daysOnZillow,
+                        'lastUpdated': lastUpdated,
+                        'onMarket': True,
                         'url': property_url,
                         'title': title}
                 properties_list.append(data)
@@ -130,7 +125,6 @@ def get_data_from_json(raw_json_data):
     except ValueError:
         print("Invalid json")
         return None
-
 
 def parse(zipcode, filter=None):
     url = create_url(zipcode, filter)
@@ -145,56 +139,14 @@ def parse(zipcode, filter=None):
     webpage = urlopen(req).read()
 
     #replace the parser to take input added above
-    #parser = html.fromstring(response.text)
     parser = html.fromstring(webpage)
-
-    # parser = html.fromstring(response.text)
     search_results = parser.xpath("//div[@id='search-results']//article")
     
-
     if not search_results:
         print("parsing from json data")
         # identified as type 2 page
         raw_json_data = parser.xpath('//script[@data-zrr-shared-data-key="mobileSearchPageStore"]//text()')
         return get_data_from_json(raw_json_data)
-
-    # print("parsing from html page")
-    # properties_list = []
-    # for properties in search_results:
-    #     raw_address = properties.xpath(".//span[@itemprop='address']//span[@itemprop='streetAddress']//text()")
-    #     raw_city = properties.xpath(".//span[@itemprop='address']//span[@itemprop='addressLocality']//text()")
-    #     raw_state = properties.xpath(".//span[@itemprop='address']//span[@itemprop='addressRegion']//text()")
-    #     raw_postal_code = properties.xpath(".//span[@itemprop='address']//span[@itemprop='postalCode']//text()")
-    #     raw_price = properties.xpath(".//span[@class='zsg-photo-card-price']//text()")
-    #     raw_info = properties.xpath(".//span[@class='zsg-photo-card-info']//text()")
-    #     raw_daysOnZillow = properties.xpath(".//span[@class='zsg-photo-card-broker-name']//text()")
-    #     url = properties.xpath(".//a[contains(@class,'overlay-link')]/@href")
-    #     raw_title = properties.xpath(".//h4//text()")
-
-    #     address = clean(raw_address)
-    #     city = clean(raw_city)
-    #     state = clean(raw_state)
-    #     postal_code = clean(raw_postal_code)
-    #     price = clean(raw_price)
-    #     info = clean(raw_info).replace(u"\xb7", ',')
-    #     daysOnZillow = clean(raw_daysOnZillow)
-    #     title = clean(raw_title)
-    #     property_url = "https://www.zillow.com" + url[0] if url else None
-    #     # is_forsale = properties.xpath('.//span[@class="zsg-icon-for-sale"]')
-
-    #     properties = {'address': address,
-    #                   'city': city,
-    #                   'state': state,
-    #                   'postal_code': postal_code,
-    #                   'price': price,
-    #                   'facts and features': info,
-    #                   'daysOnZillow': daysOnZillow,
-    #                   'url': property_url,
-    #                   'title': title}
-    #     # if is_forsale:
-    #     properties_list.append(properties)
-    # return properties_list
-
 
 if __name__ == "__main__":
     # Reading arguments
@@ -214,6 +166,5 @@ if __name__ == "__main__":
     print ("Fetching data for %s" % (zipcode))
     scraped_data = parse(zipcode, sort)
     if scraped_data:
-        print ("Writing new data to database")
-        # write_data_to_csv(scraped_data)
-        write_to_db(scraped_data, zipcode)
+        print ("Writing data to database")
+        update_db(scraped_data, zipcode)
